backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Cyber Nova <saei07@naver.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
google-generativeai = "^0.8.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"